chore: Update prompt templating to utilize role-play api
This way we will make a clear separation between input and expected output for the model.
Based on my tests Clause models struggle to read complex rules without that and treat input chunks as examples of changes.

In addition to that lang prompt moved to a system one.

metada removed as it wan't used anywhere.